Changed GyroTurn angle parameter from type double to type float

+++ FRC2018/src/PenguinEmpire.h
@@ -81,7 +81,7 @@ public:
 	void SetRightSpeed(float speed);
 	void StopMotors();
 	void TankDrive();
-	void GyroTurn(bool btn, float speed, double angle);
+	void GyroTurn(bool btn, float speed, float angle);
 	void GyroTurn(int pov, float speed);
 	void ManualShiftGears(bool upBtn, bool downBtn);
